+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+* add `subtitle` property to `Box` widget.
+* fix 1.00 pixel overflow issue by adding a `FittedBox` widget to `Box` widget.
+
 ## 1.0.0
 
 * First Release.